+++ com.microsoft.mrtk.spatialmanipulation/BoundsControl/Visuals/SqueezableBoxVisuals.cs
@@ -263,6 +263,9 @@ namespace Microsoft.MixedReality.Toolkit.SpatialManipulation
 
         protected virtual void OnEnable()
         {
+            // Ensures that when the GameObject is re-enabled, the handles are occluded ny default.
+            // If the handles are active, they will be un-occluded by the next frame.
+            // Prevents undesirable behaviour from the handle animations.
             foreach (var handle in handles)
             {
                 handle.HideOnStartup();
